pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "GalaxyTools"
version = "0.1.0"
description = "A amazing tool for restful api development"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "yifan", email = "yifan_rr@163.com"},
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "python-dotenv",
    "openai",
    "requests"
]

[project.urls]
Homepage = "https://github.com/Letcise/GalaxyTools"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"